+++ tests/dummy/app/JsonApi/Users/Schema.php
@@ -17,9 +17,10 @@
 
 namespace DummyApp\JsonApi\Users;
 
-use CloudCreativity\LaravelJsonApi\Eloquent\AbstractSchema;
+use DummyApp\User;
+use Neomerx\JsonApi\Schema\SchemaProvider;
 
-class Schema extends AbstractSchema
+class Schema extends SchemaProvider
 {
 
     /**
@@ -28,13 +29,45 @@ class Schema extends AbstractSchema
     protected $resourceType = 'users';
 
     /**
-     * @var array
+     * @param User $resource
+     * @return string
      */
-    protected $attributes = ['name', 'email'];
+    public function getId($resource)
+    {
+        return (string) $resource->getRouteKey();
+    }
 
     /**
-     * @var array
+     * @param User $resource
+     * @return array
      */
-    protected $relationships = ['phone'];
+    public function getAttributes($resource)
+    {
+        return [
+            'created-at' => $resource->created_at->toAtomString(),
+            'email' => $resource->email,
+            'name' => $resource->name,
+            'updated-at' => $resource->updated_at->toAtomString(),
+        ];
+    }
 
+    /**
+     * @param User $resource
+     * @param bool $isPrimary
+     * @param array $includeRelationships
+     * @return array
+     */
+    public function getRelationships($resource, $isPrimary, array $includeRelationships)
+    {
+        return [
+            'phone' => [
+                self::SHOW_SELF => true,
+                self::SHOW_RELATED => true,
+                self::SHOW_DATA => isset($includeRelationships['phone']),
+                self::DATA => function () use ($resource) {
+                    return $resource->phone;
+                },
+            ],
+        ];
+    }
 }